+++ MVC/Services/Interfaces/IChatService.cs
@@ -4,7 +4,7 @@ namespace Services.Interfaces
 {
     public interface IChatService
     {
-        Task<bool> AddChat(int senderId, int requestId, string message, short type);
+        Task<bool> AddChat(int senderId, int requestId, string message, int type);
 
         List<ChatMessage> GetChat(int aspNetUserId, int requestId, int type);
     }
